Refaktorerer saksoversikt sin reducer-funksjonalitet
For å få mer gjennomgående struktur på redux-filene våre, ettersom det er mye som går igjen
Følger "Ducks"-prinsippet:
https://medium.freecodecamp.org/scaling-your-redux-app-with-ducks-6115955638be
